Address encoder prefetch review feedback
