+++ megatron/core/models/mimo/model/base.py
@@ -456,7 +456,7 @@ class MimoModel(MegatronModule):
             submodule = self.modality_submodules[encoder_name]
             encoder_inputs = modality_inputs.get(encoder_name) if modality_inputs else None
             hidden_states = input_tensors.get(encoder_name) if input_tensors else None
-            output = submodule(encoder_inputs=encoder_inputs, hidden_states=hidden_states)
+            output = submodule.forward(encoder_inputs=encoder_inputs, hidden_states=hidden_states)
             if output is None and encoder_inputs is None and hidden_states is None:
                 if self._has_encoder_tokens(input_ids, encoder_name):
                     raise RuntimeError(
@@ -760,7 +760,7 @@ class MimoModel(MegatronModule):
                 and modality_inputs[modality_name] is not None
             ):
                 logger.debug(f"Processing {modality_name} modality")
-                embeddings = submodule(encoder_inputs=modality_inputs[modality_name])
+                embeddings = submodule.forward(encoder_inputs=modality_inputs[modality_name])
                 if embeddings is not None:
                     modality_embeddings[modality_name] = embeddings
                     logger.debug(
