+++ tests/unit_tests/models/mimo/test_mimo_forward_step.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 import pytest
 import torch
 
-from examples.mimo.training.encoder_prefetch import move_batch_to_cuda
+from examples.mimo.training.batch import move_batch_to_cuda
 from examples.mimo.training.step import loss_func
 from megatron.core.packed_seq_params import PackedSeqParams
